requirements.txt
aiohttp==3.7.4.post0
anyio==3.4.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
discord==1.7.3
discord.py==1.7.3
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
multidict==5.2.0
polygon==0.9.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
typing_extensions==4.0.0
urllib3==1.26.7
websocket-client==1.2.1
websockets==10.1
yarl==1.7.2
